Convert drag and drop implementation to OO
